RollingBarrel_update: (s16) cast on rotY store drops spurious extsh
The (s32) cast on the float->rotY store emitted lwz;extsh;sth; (s16) matches
target's direct lwz;sth low-half store (98.01->98.33%).

+++ src/main/dll/dll_017D_rollingbarrel.c
@@ -236,7 +236,7 @@ void RollingBarrel_update(int obj)
             }
 
             ((GameObject*)obj)->anim.rotY =
-                (s32)(lbl_803E44AC * timeDelta * state->curveSpeed +
+                (s16)(lbl_803E44AC * timeDelta * state->curveSpeed +
                     (f32)(int)((GameObject*)obj)->anim.rotY);
             hitResult = ObjHits_GetPriorityHit(obj, &hitInfo, &hitB, &hitC);
